orbweaver/requirements.txt
networkx~=2.5
PyYAML~=5.3.1
Flask~=1.1.2
google~=3.0.0
matplotlib~=3.3.3
google-api-python-client
grpcio
grpcio-tools